Implement "Choose on map" for source and destination
User can "Choose location on map". When user chooses this location, the
particular marker becomer dragable and a text is displayed asking user
to tap on map to add marker.

implemented mMap.setOnMapClickListener

When "Choose on map" is selected, focus is cleared from edittext and
keyboard is also closed.

if map is clicked, and "choose on map" option is not active, close
control panel.

Animate camera only when marker is placed through edittext