Update MapsActivity, use LatLngToPlaceTask
On using "Choose on map" or "Use my location" option, the marker was
placed on the map with a default title of "Source" or "Destination".

using LatLngToPlaceTask, we can get the name of the place from the
coordinates, then the title of the respective marker is changed, and the
entry in the history db is updated, if needed.